chore: update guardrail handler
