+++ plugins/bedrock/index.ts
@@ -1,111 +1,14 @@
-import { PluginHandler } from '../types';
-import { getText, HttpError, post } from '../utils';
-import { generateAWSHeaders } from './util';
+import { HookEventType, PluginContext, PluginHandler } from '../types';
+import {
+  getCurrentContentPart,
+  getText,
+  HttpError,
+  setCurrentContentPart,
+} from '../utils';
+import { BedrockBody, BedrockParameters } from './type';
+import { bedrockPost, redactPii } from './util';
 
 const REQUIRED_CREDENTIAL_KEYS = ['accessKeyId', 'accessKeySecret', 'region'];
-
-type BedrockFunction = 'contentFilter' | 'pii' | 'wordFilter';
-export type BedrockBody = {
-  source: 'INPUT' | 'OUTPUT';
-  content: { text: { text: string } }[];
-};
-type PIIType =
-  | 'ADDRESS'
-  | 'AGE'
-  | 'AWS_ACCESS_KEY'
-  | 'AWS_SECRET_KEY'
-  | 'CA_HEALTH_NUMBER'
-  | 'CA_SOCIAL_INSURANCE_NUMBER'
-  | 'CREDIT_DEBIT_CARD_CVV'
-  | 'CREDIT_DEBIT_CARD_EXPIRY'
-  | 'CREDIT_DEBIT_CARD_NUMBER'
-  | 'DRIVER_ID'
-  | 'EMAIL'
-  | 'INTERNATIONAL_BANK_ACCOUNT_NUMBER'
-  | 'IP_ADDRESS'
-  | 'LICENSE_PLATE'
-  | 'MAC_ADDRESS'
-  | 'NAME'
-  | 'PASSWORD'
-  | 'PHONE'
-  | 'PIN'
-  | 'SWIFT_CODE'
-  | 'UK_NATIONAL_HEALTH_SERVICE_NUMBER'
-  | 'UK_NATIONAL_INSURANCE_NUMBER'
-  | 'UK_UNIQUE_TAXPAYER_REFERENCE_NUMBER'
-  | 'URL'
-  | 'USERNAME'
-  | 'US_BANK_ACCOUNT_NUMBER'
-  | 'US_BANK_ROUTING_NUMBER'
-  | 'US_INDIVIDUAL_TAX_IDENTIFICATION_NUMBER'
-  | 'US_PASSPORT_NUMBER'
-  | 'US_SOCIAL_SECURITY_NUMBER'
-  | 'VEHICLE_IDENTIFICATION_NUMBER';
-
-interface BedrockAction<T = string> {
-  action: 'BLOCKED' | T;
-}
-
-interface ContentPolicy extends BedrockAction {
-  confidence: 'LOW' | 'NONE' | 'MEDIUM' | 'HIGH';
-  type:
-    | 'INSULTS'
-    | 'HATE'
-    | 'SEXUAL'
-    | 'VIOLENCE'
-    | 'MISCONDUCT'
-    | 'PROMPT_ATTACK';
-  filterStrength: 'LOW' | 'MEDIUM' | 'HIGH';
-}
-
-interface WordPolicy extends BedrockAction {
-  match: string;
-}
-
-interface PIIFilter extends BedrockAction<'ANONYMIZED'> {
-  match: string;
-  type: PIIType;
-}
-
-interface BedrockResponse {
-  action: 'NONE' | 'GUARDRAIL_INTERVENED';
-  assessments: {
-    wordPolicy: {
-      customWords: WordPolicy[];
-      managedWordLists: (WordPolicy & { type: 'PROFANITY' })[];
-    };
-    contentPolicy: { filters: ContentPolicy[] };
-    sensitiveInformationPolicy: {
-      piiEntities: PIIFilter[];
-      regexes: (Omit<PIIFilter, 'type'> & { name: string; regex: string })[];
-    };
-  }[];
-  output: {
-    text: string;
-  }[];
-  usage: {
-    contentPolicyUnits: number;
-    sensitiveInformationPolicyUnits: number;
-    wordPolicyUnits: number;
-  };
-}
-
-export interface BedrockParameters {
-  credentials: {
-    accessKeyId: string;
-    accessKeySecret: string;
-    awsSessionToken?: string;
-    region: string;
-  };
-  guardrailVersion: string;
-  guardrailId: string;
-}
-
-enum ResponseKey {
-  pii = 'sensitiveInformationPolicy',
-  contentFilter = 'contentPolicy',
-  wordFilter = 'wordPolicy',
-}
 
 export const validateCreds = (
   credentials?: BedrockParameters['credentials']
@@ -115,102 +18,107 @@ export const validateCreds = (
   );
 };
 
-export const bedrockPost = async (
-  credentials: Record<string, string>,
-  body: BedrockBody
-) => {
-  const url = `https://bedrock-runtime.${credentials?.region}.amazonaws.com/guardrail/${credentials?.guardrailId}/version/${credentials?.guardrailVersion}/apply`;
-
-  const headers = await generateAWSHeaders(
-    body,
-    {
-      'Content-Type': 'application/json',
-    },
-    url,
-    'POST',
-    'bedrock',
-    credentials?.region ?? 'us-east-1',
-    credentials?.accessKeyId!,
-    credentials?.accessKeySecret!,
-    credentials?.awsSessionToken || ''
-  );
-
-  return await post<BedrockResponse>(url, body, {
-    headers,
-    method: 'POST',
-  });
+const transformedData = {
+  request: {
+    json: null,
+  },
+  response: {
+    json: null,
+  },
 };
 
-export const pluginHandler: PluginHandler<BedrockParameters['credentials']> =
-  async function (
-    this: { fn: BedrockFunction },
-    context,
-    parameters,
-    eventType
-  ) {
-    const credentials = parameters.credentials;
+const handleRedaction = async (
+  context: PluginContext,
+  hookType: HookEventType,
+  credentials: Record<string, string>
+) => {
+  const { content, textArray } = getCurrentContentPart(context, hookType);
 
-    const validate = validateCreds(credentials);
+  if (!content) {
+    return [];
+  }
+  const redactPromises = textArray.map(async (text) => {
+    const result = await redactPii(text, hookType, credentials);
 
-    const guardrailVersion = parameters.guardrailVersion;
-    const guardrailId = parameters.guardrailId;
-
-    let verdict = true;
-    let error = null;
-    let data = null;
-
-    if (!validate || !guardrailVersion || !guardrailId) {
-      return {
-        verdict,
-        error: 'Missing required credentials',
-        data,
-      };
+    if (result) {
+      setCurrentContentPart(context, hookType, transformedData, result);
     }
+  });
 
-    const body = {} as BedrockBody;
+  await Promise.all(redactPromises);
+};
 
-    if (eventType === 'beforeRequestHook') {
-      body.source = 'INPUT';
-    } else {
-      body.source = 'OUTPUT';
-    }
+export const pluginHandler: PluginHandler<
+  BedrockParameters['credentials']
+> = async (context, parameters, eventType) => {
+  const credentials = parameters.credentials;
 
-    body.content = [
-      {
-        text: {
-          text: getText(context, eventType),
-        },
-      },
-    ];
+  const validate = validateCreds(credentials);
 
-    try {
-      const response = await bedrockPost(
-        { ...(credentials as any), guardrailId, guardrailVersion },
-        body
-      );
-      if (response.action === 'GUARDRAIL_INTERVENED') {
-        data = response.assessments[0]?.[ResponseKey[this.fn]];
-        if (this.fn === 'pii' && !!data) {
-          verdict = false;
-        }
-        if (this.fn === 'contentFilter' && !!data) {
-          verdict = false;
-        }
+  const guardrailVersion = parameters.guardrailVersion;
+  const guardrailId = parameters.guardrailId;
+  const pii = parameters?.piiCheck as boolean;
 
-        if (this.fn === 'wordFilter' && !!data) {
-          verdict = false;
-        }
-      }
-    } catch (e) {
-      if (e instanceof HttpError) {
-        error = e.response.body;
-      } else {
-        error = (e as Error).message;
-      }
-    }
+  let verdict = true;
+  let error = null;
+  let data = null;
+  if (!validate || !guardrailVersion || !guardrailId) {
     return {
       verdict,
-      error,
+      error: 'Missing required credentials',
       data,
     };
+  }
+
+  if (pii) {
+    await handleRedaction(context, eventType, {
+      ...credentials,
+      guardrailId,
+      guardrailVersion,
+    });
+
+    return { error, data, verdict: true, transformedData };
+  }
+
+  const body = {} as BedrockBody;
+
+  if (eventType === 'beforeRequestHook') {
+    body.source = 'INPUT';
+  } else {
+    body.source = 'OUTPUT';
+  }
+
+  body.content = [
+    {
+      text: {
+        text: getText(context, eventType),
+      },
+    },
+  ];
+
+  try {
+    const response = await bedrockPost(
+      { ...(credentials as any), guardrailId, guardrailVersion },
+      body
+    );
+    if (response.action === 'GUARDRAIL_INTERVENED') {
+      verdict = false;
+      // Send assessments
+      data = response.assessments[0] as any;
+
+      delete data['invocationMetrics'];
+      delete data['usage'];
+    }
+  } catch (e) {
+    if (e instanceof HttpError) {
+      error = e.response.body;
+    } else {
+      error = (e as Error).message;
+    }
+  }
+  return {
+    verdict,
+    error,
+    data,
   };
+};
