+++ plugins/bedrock/manifest.json
@@ -32,8 +32,8 @@
 
   "functions": [
     {
-      "name": "Content Filter",
-      "id": "contentFilter",
+      "name": "Apply Bedrock guardrail",
+      "id": "guard",
       "type": "guardrail",
       "supportedHooks": ["beforeRequestHook", "afterRequestHook"],
       "description": [
@@ -54,34 +54,15 @@
             "type": "string",
             "label": "Guardrail ID",
             "description": "ID of the guardrail."
+          },
+          "piiCheck": {
+            "type": "boolean",
+            "label": "PII Guard",
+            "description": "Enable Personally Identifiable Information(PII) check"
           }
         },
         "required": ["guardrailVersion", "guardrailId"]
       }
-    },
-    {
-      "id": "pii",
-      "name": "Sensitive Content",
-      "type": "guardrail",
-      "supportedHooks": ["beforeRequestHook", "afterRequestHook"],
-      "description": [
-        {
-          "type": "subHeading",
-          "text": "Checks if the content contains any Personally Identifiable Information (PII)."
-        }
-      ]
-    },
-    {
-      "id": "wordFilter",
-      "name": "Word Filter",
-      "type": "guardrail",
-      "supportedHooks": ["beforeRequestHook", "afterRequestHook"],
-      "description": [
-        {
-          "type": "subHeading",
-          "text": "Filters out words that are not allowed in the content."
-        }
-      ]
     }
   ]
 }
