+++ plugins/index.ts
@@ -42,7 +42,6 @@ import { handler as pangearedactPii } from './pangea/redactPii';
 import { handler as patronusredactPii } from './patronus/redactPii';
 import { handler as patronusredactPhi } from './patronus/redactPhi';
 import { pluginHandler as bedrockHandler } from './bedrock/index';
-import { bedrockPIIHandler } from './bedrock/redactPii';
 
 export const plugins = {
   default: {
@@ -107,9 +106,6 @@ export const plugins = {
     guard: promptfooGuard,
   },
   bedrock: {
-    pii: bedrockHandler.bind({ fn: 'pii' }),
-    contentFilter: bedrockHandler.bind({ fn: 'contentFilter' }),
-    wordFilter: bedrockHandler.bind({ fn: 'wordFilter' }),
-    redactPii: bedrockPIIHandler,
+    guard: bedrockHandler,
   },
 };
